this/requirements.txt
blinker==1.7.0
boto3==1.34.0
botocore==1.34.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
greenlet==3.0.2
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mod_wsgi==5.0.0
mysqlclient==2.2.0
pillow==10.2.0
pyotp==2.9.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0